pyproject.toml
[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[project]
name = "arioso"
version = "0.0.3"
description = "Make tunes"
readme = "README.md"
requires-python = ">=3.10"
keywords = []
authors = [
    { name = "Thor Whalen" },
]
dependencies = []

[project.license]
text = "mit"

[project.urls]
Homepage = "https://github.com/thorwhalen/arioso"
Repository = "https://github.com/thorwhalen/arioso"
Documentation = "https://thorwhalen.github.io/arioso"

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "numpy",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "numpy",
    "ruff>=0.1.0",
]
docs = [
    "sphinx>=6.0",
    "sphinx-rtd-theme>=1.0",
]
musicgen = [
    "audiocraft>=1.0",
    "torch>=2.0",
]
musicgen-transformers = [
    "transformers>=4.30",
    "torch>=2.0",
]
sunoapi = [
    "requests>=2.28",
]
elevenlabs = [
    "requests>=2.28",
    "ho",
    "ju",
]
stable-audio = [
    "diffusers>=0.25",
    "torch>=2.0",
    "transformers>=4.30",
]
lyria2 = [
    "requests>=2.28",
    "google-cloud-aiplatform>=1.38",
]
lyria-rt = [
    "google-genai>=1.0",
]
mubert = [
    "requests>=2.28",
]
beatoven = [
    "requests>=2.28",
]
loudly = [
    "requests>=2.28",
]
jen = [
    "requests>=2.28",
]
yue = [
    "fal-client>=0.4",
]
harmonai = [
    "diffusers>=0.25",
    "torch>=2.0",
]
riffusion = [
    "diffusers>=0.25",
    "torch>=2.0",
]
udio = [
    "udio-wrapper",
]
all-rest = [
    "requests>=2.28",
    "ho",
    "ju",
]
all = [
    "audiocraft>=1.0",
    "torch>=2.0",
    "requests>=2.28",
    "ho",
    "ju",
    "diffusers>=0.25",
    "transformers>=4.30",
    "google-genai>=1.0",
    "fal-client>=0.4",
]

[tool.ruff]
line-length = 88
target-version = "py310"
exclude = [
    "**/*.ipynb",
    ".git",
    ".venv",
    "build",
    "dist",
    "tests",
    "examples",
    "scrap",
]

[tool.ruff.lint]
select = [
    "D100",
]
ignore = [
    "D203",
    "E501",
    "B905",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"**/tests/*" = [
    "D",
]
"**/examples/*" = [
    "D",
]
"**/scrap/*" = [
    "D",
]

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests",
]
doctest_optionflags = [
    "NORMALIZE_WHITESPACE",
    "ELLIPSIS",
]

[tool.wads.ci]
project_name = ""

[tool.wads.ci.commands]
pre_test = []
test = []
post_test = []
lint = []
format = []

[tool.wads.ci.env]
required_envvars = []
test_envvars = []
extra_envvars = []

[tool.wads.ci.env.defaults]

[tool.wads.ci.quality.ruff]
enabled = true

[tool.wads.ci.quality.black]
enabled = false

[tool.wads.ci.quality.mypy]
enabled = false

[tool.wads.ci.testing]
python_versions = [
    "3.10",
    "3.12",
]
pytest_args = [
    "-v",
    "--tb=short",
    "tests",
]
skip_doctests = true
coverage_enabled = true
coverage_threshold = 0
coverage_report_format = [
    "term",
    "xml",
]
exclude_paths = [
    "examples",
    "scrap",
]
test_on_windows = true

[tool.wads.ci.metrics]
enabled = true
config_path = ".github/umpyre-config.yml"
storage_branch = "code-metrics"
python_version = "3.10"
force_run = false

[tool.wads.ci.build]
sdist = true
wheel = true

[tool.wads.ci.publish]
enabled = true

[tool.wads.ci.docs]
enabled = true
builder = "epythet"
ignore_paths = [
    "tests/",
    "scrap/",
    "examples/",
]
